# docs/openapi-short-url.yaml
openapi: 3.1.0
info:
  title: Swoole Short URL Service API
  version: 1.2.0
  description: |
    Production-like short URL API specification with async visit log worker,
    idempotent create, and admin-protected management endpoints.
servers:
  - url: http://127.0.0.1:9501
paths:
...
  /api/v1/short-urls:
    post:
      summary: Create short URL
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
                custom_code:
                  type: string
                expires_at:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateShortUrlResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/v1/admin/short-urls:
    get:
      summary: List short URLs
      security:
        - AdminApiKey: []
      responses:
        '200':
          description: Listed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminShortUrlListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  securitySchemes:
    AdminApiKey:
      type: apiKey
      in: header
      name: X-Admin-Api-Key
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, trace_id]
          properties:
            code:
              type: string
            message:
              type: string
            trace_id:
              type: string
            details:
              type: object
              additionalProperties: true
...
    ShortUrlData:
      type: object
      required: [code, short_url, original_url, created_at, is_active]
      properties:
        code:
          type: string
        short_url:
          type: string
        original_url:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
        expires_at:
          type: [string, 'null']
          format: date-time
        is_active:
          type: boolean
    CreateShortUrlResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/ShortUrlData'
    AdminShortUrlListItem:
      type: object
      required:
        [code, short_url, original_url, is_active, total_visits, created_at]
      properties:
        code:
          type: string
        short_url:
          type: string
        original_url:
          type: string
          format: uri
        is_active:
          type: boolean
        total_visits:
          type: integer
        created_at:
          type: string
          format: date-time
        expires_at:
          type: [string, 'null']
          format: date-time
        last_visited_at:
          type: [string, 'null']
          format: date-time
    AdminShortUrlListData:
      type: object
      required: [items, page, per_page, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminShortUrlListItem'
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
    AdminShortUrlListResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/AdminShortUrlListData'
...